+++ crnd_wsd/tests/test_tour.py
@@ -15,6 +15,21 @@ class TestWebsiteServiceDesk(TestPhantomTour):
         self.user_demo = self.env.ref(
             'crnd_wsd.user_demo_service_desk_website')
         self.group_portal = self.env.ref('base.group_portal')
+        self.user_demo = self.env.ref(
+            'crnd_wsd.user_demo_service_desk_website')
+        self.group_portal = self.env.ref('base.group_portal')
+
+        # Tests response attachments case
+        self.request_resp_attachments = self.env.ref(
+            'generic_request.request_request_type_simple_demo_1')
+        self.response_attachment1 = self.env.ref(
+            'generic_request.request_response_attachment_demo1')
+        self.response_attachment2 = self.env.ref(
+            'generic_request.request_response_attachment_demo2')
+        self.stage_sent = self.env.ref(
+            'generic_request.request_stage_type_simple_sent')
+        self.stage_confirmed = self.env.ref(
+            'generic_request.request_stage_type_simple_confirmed')
 
     def test_tour_request_base(self):
         self.assertIn(self.group_portal, self.user_demo.groups_id)
@@ -77,6 +92,8 @@ class TestWebsiteServiceDesk(TestPhantomTour):
 
         self.env['request.request']._patch_method('create', monkey_create)
 
+        self._disable_use_services_setting()
+
         new_requests = self._test_phantom_tour_requests(
             '/', 'crnd_wsd_tour_request_new',
             login=self.user_demo.login)
@@ -98,17 +115,31 @@ class TestWebsiteServiceDesk(TestPhantomTour):
         ).write({
             'default_request_text': 'Default text',
         })
-        self._test_phantom_tour(
+
+        self._disable_use_services_setting()
+
+        request = self._test_phantom_tour_requests(
             '/', 'crnd_wsd_tour_request_new_default_text',
             login=self.user_demo.login)
 
+        # Check channel 'Website' is assigned to the request
+        channel_website = self.env.ref(
+            'generic_request.request_channel_website')
+        self.assertEqual(request.channel_id, channel_website)
+
     def test_tour_public_user(self):
         self.env.user.company_id.request_wsd_public_ui_visibility = 'restrict'
+
+        self._disable_use_services_setting()
+
         self._test_phantom_tour(
             '/', 'crnd_wsd_tour_request_public_user')
 
     def test_tour_public_user_redirect(self):
         self.env.user.company_id.request_wsd_public_ui_visibility = 'redirect'
+
+        self._disable_use_services_setting()
+
         self._test_phantom_tour(
             '/', 'crnd_wsd_tour_request_public_user_redirect')
 
@@ -120,6 +151,8 @@ class TestWebsiteServiceDesk(TestPhantomTour):
             default_website.request_redirect_after_created_on_website,
             'congrats_page'
         )
+
+        self._disable_use_services_setting()
 
         request = self._test_phantom_tour_requests(
             '/',
@@ -149,6 +182,8 @@ class TestWebsiteServiceDesk(TestPhantomTour):
             'req_page'
         )
 
+        self._disable_use_services_setting()
+
         request = self._test_phantom_tour_requests(
             '/',
             'crnd_wsd_tour_request_public_user_create_req_to_request')
@@ -165,6 +200,9 @@ class TestWebsiteServiceDesk(TestPhantomTour):
             'create-request')
         company = self.env.user.company_id
         company.request_mail_create_author_contact_from_email = True
+
+        self._disable_use_services_setting()
+
         request = self._test_phantom_tour_requests(
             '/',
             'crnd_wsd_tour_request_public_user_create_req_to_congrat')
@@ -183,6 +221,9 @@ class TestWebsiteServiceDesk(TestPhantomTour):
             'create-request')
         self.env.user.company_id.request_wsd_public_use_author_phone = (
             'no-phone')
+
+        self._disable_use_services_setting()
+
         request = self._test_phantom_tour_requests(
             '/',
             'crnd_wsd_tour_request_author_no_phone')
@@ -194,6 +235,9 @@ class TestWebsiteServiceDesk(TestPhantomTour):
             'create-request')
         self.env.user.company_id.request_wsd_public_use_author_phone = (
             'required-phone')
+
+        self._disable_use_services_setting()
+
         request = self._test_phantom_tour_requests(
             '/',
             'crnd_wsd_tour_request_author_phone_required')
@@ -206,6 +250,9 @@ class TestWebsiteServiceDesk(TestPhantomTour):
             'create-request')
         self.env.user.company_id.request_wsd_public_use_author_phone = (
             'optional-phone')
+
+        self._disable_use_services_setting()
+
         request_without_phone = self._test_phantom_tour_requests(
             '/',
             'crnd_wsd_tour_request_author_no_phone')
@@ -219,9 +266,212 @@ class TestWebsiteServiceDesk(TestPhantomTour):
             'create-request')
         self.env.user.company_id.request_wsd_public_use_author_phone = (
             'optional-phone')
+
+        self._disable_use_services_setting()
+
         request_with_phone = self._test_phantom_tour_requests(
             '/',
             'crnd_wsd_tour_request_author_phone_required')
         self.assertEqual(request_with_phone.email_from,
                          'Test_author@email.com')
         self.assertEqual(request_with_phone.author_phone, '123456789')
+
+    def test_tour_request_channel_website_archived(self):
+        channel_other = self.env.ref('generic_request.request_channel_other')
+
+        # Archive channel website
+        channel_website = self.env.ref(
+            'generic_request.request_channel_website')
+        channel_website.sudo().active = False
+        channel_website.invalidate_cache()
+
+        # Create Website request
+        self.env.ref(
+            'crnd_wsd.request_type_generic'
+        ).write({
+            'default_request_text': 'Default text',
+        })
+
+        self._disable_use_services_setting()
+
+        request = self._test_phantom_tour_requests(
+            '/', 'crnd_wsd_tour_request_new_default_text',
+            login=self.user_demo.login)
+
+        # Check that default channel assigned to the request
+        self.assertEqual(request.channel_id, channel_other)
+
+    def test_tour_request_channel_website_deleted(self):
+        channel_other = self.env.ref('generic_request.request_channel_other')
+
+        # Delete channel website
+        self.env.ref(
+            'generic_request.request_channel_website').sudo().unlink()
+
+        # Create Website request
+        self.env.ref(
+            'crnd_wsd.request_type_generic'
+        ).write({
+            'default_request_text': 'Default text',
+        })
+
+        self._disable_use_services_setting()
+
+        request = self._test_phantom_tour_requests(
+            '/', 'crnd_wsd_tour_request_new_default_text',
+            login=self.user_demo.login)
+
+        # Check that default channel assigned to the request
+        self.assertEqual(request.channel_id, channel_other)
+
+    def test_tour_request_new_priority_start(self):
+        self.env.ref(
+            'crnd_wsd.request_type_generic'
+        ).write({
+            'complex_priority': False,
+            'selection_priority_view': 'star',
+        })
+
+        self._disable_use_services_setting()
+
+        new_requests = self._test_phantom_tour_requests(
+            '/', 'crnd_wsd_tour_request_new_star',
+            login=self.user_demo.login)
+
+        self.assertEqual(len(new_requests), 1)
+        self.assertEqual(new_requests.priority, '3')
+
+    def test_tour_request_new_priority_complex_start(self):
+        self.env.ref(
+            'crnd_wsd.request_type_generic'
+        ).write({
+            'complex_priority': True,
+            'selection_priority_view': 'star',
+        })
+
+        self._disable_use_services_setting()
+
+        new_requests = self._test_phantom_tour_requests(
+            '/', 'crnd_wsd_tour_request_new_complex_star',
+            login=self.user_demo.login)
+
+        self.assertEqual(len(new_requests), 1)
+        self.assertEqual(new_requests.impact, '2')
+        self.assertEqual(new_requests.urgency, '1')
+
+    def test_tour_request_new_priority_selection(self):
+        self.env.ref(
+            'crnd_wsd.request_type_generic'
+        ).write({
+            'complex_priority': False,
+            'selection_priority_view': 'selection',
+        })
+
+        self._disable_use_services_setting()
+
+        new_requests = self._test_phantom_tour_requests(
+            '/', 'crnd_wsd_tour_request_new_selection',
+            login=self.user_demo.login)
+
+        self.assertEqual(len(new_requests), 1)
+        self.assertEqual(new_requests.priority, '4')
+
+    def test_tour_request_new_priority_complex_selection(self):
+        self.env.ref(
+            'crnd_wsd.request_type_generic'
+        ).write({
+            'complex_priority': True,
+            'selection_priority_view': 'selection',
+        })
+
+        self._disable_use_services_setting()
+
+        new_requests = self._test_phantom_tour_requests(
+            '/', 'crnd_wsd_tour_request_new_complex_selection',
+            login=self.user_demo.login)
+
+        self.assertEqual(len(new_requests), 1)
+        self.assertEqual(new_requests.impact, '1')
+        self.assertEqual(new_requests.urgency, '3')
+
+    def test_tour_crnd_wsd_subrequest(self):
+        self._disable_use_services_setting()
+
+        self._test_phantom_tour(
+            '/', 'tour_crnd_wsd_subrequest',
+            login=self.user_demo.login)
+
+    def test_tour_crnd_wsd_subrequest_additional(self):
+        self._disable_use_services_setting()
+
+        self._test_phantom_tour(
+            '/', 'tour_crnd_wsd_subrequest_additional',
+            login=self.user_demo.login)
+
+    def test_tour_request_new_service(self):
+        self._enable_use_services_setting()
+        self.assertIn(self.group_portal, self.user_demo.groups_id)
+
+        # TODO: refactor demo data to avoid using non-demo data
+        # show on website notebook rent service
+        self.env.ref('crnd_service_desk.request_type_incident').write({
+            'website_published': True,
+            'service_ids': [(4, self.env.ref(
+                'generic_service.generic_service_default').id)],
+            'category_ids': [(4, self.env.ref(
+                'crnd_wsd.'
+                'request_category_demo_website_support').id)],
+        })
+        self.env.ref('generic_service.generic_service_rent_notebook').write({
+            'request_type_ids': [
+                (4, self.env.ref(
+                    'crnd_service_desk.request_type_incident').id),
+            ],
+            'category_ids': [
+                (4, self.env.ref(
+                    'crnd_wsd.'
+                    'request_category_demo_website_support').id),
+            ],
+            'website_published': True,
+        })
+
+        new_requests = self._test_phantom_tour_requests(
+            '/', 'crnd_wsd_tour_request_new_with_service',
+            login=self.user_demo.login)
+
+        self.assertEqual(len(new_requests), 1)
+        self.assertTrue(
+            new_requests.request_text.startswith(
+                u'<h1>Test incident request</h1>'))
+        self.assertEqual(
+            new_requests.service_id,
+            self.env.ref('generic_service.generic_service_default'))
+
+    def test_tour_check_response_attachments(self):
+        # Close request and add response attachments
+        self.request_resp_attachments.write({
+            'author_id': self.env.ref('base.user_admin').id,
+            'stage_id': self.stage_sent.id,
+            'request_text': 'Test response attachments'
+        })
+        self.assertTrue(self.request_resp_attachments.can_be_closed)
+        close_route = self.env['request.stage.route'].sudo().search([
+            ('request_type_id', '=', self.request_resp_attachments.type_id.id),
+            ('stage_from_id', '=', self.request_resp_attachments.stage_id.id),
+            ('stage_to_id', '=', self.stage_confirmed.id),
+        ])
+        wizard_close = self.env['request.wizard.close'].create({
+            'request_id': self.request_resp_attachments.id,
+            'close_route_id': close_route.id,
+            'response_text': 'test-response-attachments',
+            'attachment_ids': [(4, self.response_attachment1.id),
+                               (4, self.response_attachment2.id)],
+        })
+        wizard_close.action_close_request()
+
+        # Check response attachments on website
+        self.run_js_tour(
+            "/web",
+            'tour_response_attachments',
+            login='admin',
+        )