+++ generic_request/wizard/base_partner_merge.py
@@ -6,6 +6,7 @@ class MergePartnerAutomatic(models.TransientModel):
 
     def action_merge(self):
         res = super(MergePartnerAutomatic, self).action_merge()
+        self.env['res.partner'].invalidate_cache(fnames=['request_ids'])
         self.dst_partner_id.modified(fnames=[
             'request_by_partner_ids', 'request_by_author_ids'
         ])